requirements.txt
streamlit
pandas
openai
langchain
langchain-community
langchain-openai
PyPDF2
faiss-cpu
openpyxl
PyJWT==2.8.0
